requirements_old.txt
fastapi~=0.85.0
uvicorn
jinja2
python-multipart
openai-whisper
numpy
torch
tqdm
more-itertools
transformers>=4.19.0
ffmpeg-python==0.2.0
datetime
passlib

